Move pick-all-lower-objects-numbers util into separate file
