+++ workers/loc.api/helpers/utils.js
@@ -155,36 +155,6 @@ const pickLowerObjectsNumbers = (propName, objects = []) => {
   }, null)
 }
 
-const pickAllLowerObjectsNumbers = (propName, objects = []) => {
-  return objects.reduce((accum, curr) => {
-    if (
-      !curr?.[propName] ||
-      typeof curr?.[propName] !== 'object'
-    ) {
-      return accum
-    }
-
-    const entries = Object.entries(curr[propName])
-
-    return entries.reduce((accum, [key, val]) => {
-      if (!Number.isFinite(val)) {
-        return accum
-      }
-      if (!Number.isFinite(accum?.[key])) {
-        accum[key] = val
-
-        return accum
-      }
-
-      accum[key] = val < accum[key]
-        ? val
-        : accum[key]
-
-      return accum
-    }, accum)
-  }, {})
-}
-
 module.exports = {
   checkParamsAuth,
   tryParseJSON,
@@ -192,6 +162,5 @@ module.exports = {
   getDateString,
   isNotSyncRequired,
   sumObjectsNumbers,
-  pickLowerObjectsNumbers,
-  pickAllLowerObjectsNumbers
+  pickLowerObjectsNumbers
 }
