+++ workers/loc.api/helpers/utils/index.js
@@ -4,8 +4,10 @@ const orderBy = require('./order-by')
 const pushLargeArr = require('./push-large-arr')
 const sumArrayVolumes = require('./sum-array-volumes')
 const sumAllObjectsNumbers = require('./sum-all-objects-numbers')
+const pickAllLowerObjectsNumbers = require('./pick-all-lower-objects-numbers')
 
 module.exports = {
+  pickAllLowerObjectsNumbers,
   sumAllObjectsNumbers,
   sumArrayVolumes,
   pushLargeArr,
